[1,2,3 단계]Web MVC  (#8)
* feat: 로그인(토큰 발급) 및 내 정보 조회 기능 구현

* feat: 예약시 혹은 예약 취소시 로그인 유저인지 확인

* feat: admin 인터셉터 구현

+++ src/main/java/nextstep/reservation/ReservationService.java
@@ -1,5 +1,7 @@
 package nextstep.reservation;
 
+import nextstep.member.MemberDao;
+import nextstep.reservation.presentation.dto.ReservationRequest;
 import nextstep.schedule.Schedule;
 import nextstep.schedule.ScheduleDao;
 import nextstep.support.DuplicateEntityException;
@@ -11,17 +13,19 @@ import java.util.List;
 
 @Service
 public class ReservationService {
-    public final ReservationDao reservationDao;
-    public final ThemeDao themeDao;
-    public final ScheduleDao scheduleDao;
+    private final ReservationDao reservationDao;
+    private final ThemeDao themeDao;
+    private final ScheduleDao scheduleDao;
+    private final MemberDao memberDao;
 
-    public ReservationService(ReservationDao reservationDao, ThemeDao themeDao, ScheduleDao scheduleDao) {
+    public ReservationService(ReservationDao reservationDao, ThemeDao themeDao, ScheduleDao scheduleDao, MemberDao memberDao) {
         this.reservationDao = reservationDao;
         this.themeDao = themeDao;
         this.scheduleDao = scheduleDao;
+        this.memberDao = memberDao;
     }
 
-    public Long create(ReservationRequest reservationRequest) {
+    public Long create(ReservationRequest reservationRequest, String uuid) {
         Schedule schedule = scheduleDao.findById(reservationRequest.getScheduleId());
         if (schedule == null) {
             throw new NullPointerException();
@@ -32,9 +36,10 @@ public class ReservationService {
             throw new DuplicateEntityException();
         }
 
+        Long memberId = memberDao.findByUuid(uuid).orElseThrow().getId();
         Reservation newReservation = new Reservation(
                 schedule,
-                reservationRequest.getName()
+                memberId
         );
 
         return reservationDao.save(newReservation);
@@ -49,10 +54,15 @@ public class ReservationService {
         return reservationDao.findAllByThemeIdAndDate(themeId, date);
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Long id, String uuid) {
         Reservation reservation = reservationDao.findById(id);
         if (reservation == null) {
             throw new NullPointerException();
+        }
+
+        Long memberId = memberDao.findByUuid(uuid).orElseThrow().getId();
+        if(!reservation.getMemberId().equals(memberId)) {
+            throw new IllegalArgumentException("해당 회원이 취소할 수 없는 예약입니다.");
         }
 
         reservationDao.deleteById(id);
